split "chatting.py" into "contact.py" and "chat.py" to be more reminiscent of the core-rust structure
